+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## 1.5.0
 
+* Add `showNA` argument to `vectorShow()`.
 * Change `oce.plot.ts()` by adding `simplify` argument.
 * Fix `[[` to yield ITS90 temperature for all classes, not just `ctd`.
 * Fix `as.ctd()` handling of temperature scale of first argument.